requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
aiohttp>=3.8.0
pydantic>=1.8.0

# Environment Variables
python-dotenv>=0.19.0

# AI/Chat
g4f>=0.1.0

# Machine Learning & Game Environment
gym>=0.21.0
stable-retro>=0.9.0
stable-baselines3>=1.5.0
numpy>=1.19.0
torch>=1.9.0

# Development Tools
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0
mypy>=0.900

# Build Tools
build>=0.7.0
twine>=3.4.0
wheel>=0.37.0
setuptools>=65.0.0